Change hash value field to uint64 in bm_op_t and update every part which uses it. Remove bm_record_read_op and bm_record_write_op procedures and add bm_record_op instead. So the code could be used by memcached and redis.

+++ k_v_benchmark.h
@@ -19,7 +19,7 @@ typedef enum {
 
 typedef struct {
     bm_op_type_t type;
-    uint32_t	 key_hv;
+    uint64_t	 key_hv;
 } bm_op_t;
 
 #endif
